+++ functional_tests/ft_paper_example_2.R
@@ -1,15 +1,17 @@
 ##      author:  Naeem Khoshnevis
-##      created: August 2022
+##      created: October 2022
 ##      purpose: Reproducing example 2 in the paper.
 
 # Load libraries
 library(ggplot2)
 
 # CausalGPS: Matching on Generalized Propensity Scores with Continuous Exposures
-# Example 1_1: Use all data and default values.
+# Example 2: Use all data and default values.
 #
 
 data("synthetic_us_2010", package = "CausalGPS")
+
+# cdc_mean_bmi larger that 9000 is not a real data.
 synthetic_us_2010$cdc_mean_bmi[synthetic_us_2010$cdc_mean_bmi > 9000] <- NA
 data <- synthetic_us_2010
 
@@ -45,28 +47,19 @@ confounders_s1 <- c("cs_poverty","cs_hispanic",
 # data preprocessing
 data$region <- as.factor(data$region)
 
-
-## Plotting GPS distribution
-q1 <- stats::quantile(data$qd_mean_pm25,0.25)
-q2 <- stats::quantile(data$qd_mean_pm25,0.99)
-
-trimmed_data <- subset(data[stats::complete.cases(data) ,],
-                       qd_mean_pm25 <= q2  & qd_mean_pm25 >= q1)
-
 log1 <- function(x){log(abs(x)+0.001)}
 
 set.seed(369)
 set_logger(logger_level = "DEBUG")
-ps_pop_obj_1 <- generate_pseudo_pop(trimmed_data$cms_mortality_pct,
-                                    trimmed_data$qd_mean_pm25,
-                                    data.frame(trimmed_data[, confounders_s1,
+ps_pop_obj_1 <- generate_pseudo_pop(data$cms_mortality_pct,
+                                    data$qd_mean_pm25,
+                                    data.frame(data[, confounders_s1,
                                                             drop=FALSE]),
-                                    ci_appr = "matching",
-                                    pred_model = "sl",
+                                    ci_appr = "weighting",
                                     gps_model = "non-parametric",
                                     bin_seq = NULL,
-                                    trim_quantiles = c(0.0 ,
-                                                       1.0),
+                                    trim_quantiles = c(0.25 ,
+                                                       0.95),
                                     optimized_compile = TRUE,
                                     use_cov_transform = FALSE,
                                     transformers = list("pow2","pow3"),
@@ -77,135 +70,27 @@ ps_pop_obj_1 <- generate_pseudo_pop(trimmed_data$cms_mortality_pct,
                                     covar_bl_method = "absolute",
                                     covar_bl_trs = 0.1,
                                     covar_bl_trs_type= "maximal",
-                                    max_attempt = 200,
+                                    max_attempt = 10,
                                     matching_fun = "matching_l1",
                                     delta_n = 0.1,
                                     scale = 1)
 
 plot(ps_pop_obj_1)
 
-,
-rgr_num.trees=seq(10,200,1)
 
-pdf("example_2_1.pdf")
-plot(ps_pop_obj_1)
-dev.off()
-
-abc
-
-### Example 1-3: Use trimmed data and search for the best params.
-# Takes 45 iteration to get 0.040 covariate balance
-set.seed(172)
-ps_pop_obj_3 <- generate_pseudo_pop(trimmed_data$cms_mortality_pct,
-                                    trimmed_data$qd_mean_pm25,
-                                    data.frame(trimmed_data[, confounders_s1,
-                                                            drop=FALSE]),
-                                    ci_appr = "matching",
-                                    pred_model = "sl",
-                                    gps_model = "parametric",
-                                    bin_seq = NULL,
-                                    trim_quantiles = c(0.0,
-                                                       1.0),
-                                    optimized_compile = TRUE,
-                                    use_cov_transform = FALSE,
-                                    transformers = list("pow2","pow3"),
-                                    sl_lib = c("m_xgboost"),
-                                    params = list(xgb_nrounds=seq(10,50,1),
-                                                  xgb_eta=seq(0.1,0.4,0.01)),
-                                    nthread = 12,
-                                    covar_bl_method = "absolute",
-                                    covar_bl_trs = 0.1,
-                                    covar_bl_trs_type= "maximal",
-                                    max_attempt = 100,
-                                    matching_fun = "matching_l1",
-                                    delta_n = 0.1,
-                                    scale = 1)
-
-
-pdf("example_1_3.pdf")
-plot(ps_pop_obj_3)
-dev.off()
-
-summary(ps_pop_obj_3)
-
-
-### Example 1-4: Similar to 1-3, but using the optimized values.
-set.seed(172)
-ps_pop_obj_4 <- generate_pseudo_pop(trimmed_data$cms_mortality_pct,
-                                    trimmed_data$qd_mean_pm25,
-                                    data.frame(trimmed_data[, confounders_s1,
-                                                            drop=FALSE]),
-                                    ci_appr = "matching",
-                                    pred_model = "sl",
-                                    gps_model = "parametric",
-                                    bin_seq = NULL,
-                                    trim_quantiles = c(0.0,
-                                                       1.0),
-                                    optimized_compile = TRUE,
-                                    use_cov_transform = FALSE,
-                                    transformers = list("pow2","pow3"),
-                                    sl_lib = c("m_xgboost"),
-                                    params = list(xgb_nrounds=c(18),
-                                                  xgb_eta=c(0.22)),
-                                    nthread = 12,
-                                    covar_bl_method = "absolute",
-                                    covar_bl_trs = 0.1,
-                                    covar_bl_trs_type= "maximal",
-                                    max_attempt = 1,
-                                    matching_fun = "matching_l1",
-                                    delta_n = 0.1,
-                                    scale = 1)
-
-
-pdf("example_1_4.pdf")
-plot(ps_pop_obj_4)
-dev.off()
-
-summary(ps_pop_obj_4)
-
-########
-# example_1_5: Similar to the results of example_1_3 but with transformers.
-# Gets almost the same results.
-set.seed(172)
-ps_pop_obj_4 <- generate_pseudo_pop(trimmed_data$cms_mortality_pct,
-                                    trimmed_data$qd_mean_pm25,
-                                    data.frame(trimmed_data[, confounders_s1,
-                                                            drop=FALSE]),
-                                    ci_appr = "matching",
-                                    pred_model = "sl",
-                                    gps_model = "parametric",
-                                    bin_seq = NULL,
-                                    trim_quantiles = c(0.0,
-                                                       1.0),
-                                    optimized_compile = TRUE,
-                                    use_cov_transform = TRUE,
-                                    transformers = list("pow2","pow3"),
-                                    sl_lib = c("m_xgboost"),
-                                    params = list(xgb_nrounds=seq(10,50,1),
-                                                  xgb_eta=seq(0.1,0.4,0.01)),
-                                    nthread = 12,
-                                    covar_bl_method = "absolute",
-                                    covar_bl_trs = 0.06,
-                                    covar_bl_trs_type= "maximal",
-                                    max_attempt = 100,
-                                    matching_fun = "matching_l1",
-                                    delta_n = 0.1,
-                                    scale = 1)
-
-
-
-#########
-# Exposure response function for example_1_4
+# pdf("example_2_1.pdf")
+# plot(ps_pop_obj_1)
+# dev.off()
 
 set.seed(168)
-erf <- estimate_npmetric_erf(matched_Y = ps_pop_obj_4$pseudo_pop$Y,
-                             matched_w = ps_pop_obj_4$pseudo_pop$w,
-                             matched_counter = ps_pop_obj_4$pseudo_pop$counter,
+erf <- estimate_npmetric_erf(matched_Y = ps_pop_obj_1$pseudo_pop$Y,
+                             matched_w = ps_pop_obj_1$pseudo_pop$w,
+                             matched_counter = NULL,
                              bw_seq = seq(0.5,4,0.1),
                              w_vals = seq(7,13, 0.05),
                              nthread = 12)
 
-pdf("example_1_erf.pdf")
-plot(erf, gg_labs = c("PM2.5", "All-cause Mortality"),
-     gg_title = c("Exposure Response Curve"))
-dev.off()
+# pdf("example_1_erf.pdf")
+# plot(erf, gg_labs = c("PM2.5", "All-cause Mortality"),
+#      gg_title = c("Exposure Response Curve"))
+# dev.off()