+++ NEWS.md
@@ -5,6 +5,7 @@
 
 ### Changed
 
+* Internal kernel smoothing now uses `locpol::locpol` function.
 * The entire data set is trimmed based on trimming quantiles.
 * `earth` and `ranger` are not installed automatically. They can be installed manually if needed.
 * `sysdata.rda` is modified to reflect transition from `counter` and `ipw` to `counter_weight`
@@ -20,6 +21,7 @@ approaches. `counter` and `ipw` are dropped.
 
 ### Removed 
 
+* Dropped importing `KernSmooth` package. 
 * `pred_model` argument dropped. The package only predicts using SuperLearner.